requirements.txt
requests==2.27.1
dhooks==1.1.4
ujson==5.2.0
git+https://github.com/Rapptz/discord.py
python-dotenv==0.20.0
psutil==5.9.0